use double for arch linux

+++ lsd_slam_viewer/src/PointCloudViewer.h
@@ -132,8 +132,8 @@ public:
 		int showCurrentCam_i = showCurrentCam;
 		int isFix_i = isFix;
 
-		float x,y,z;
-		frame.getPosition(x,y,z);
+        double x,y,z;
+        frame.getPosition(x,y,z);
 
     	snprintf(buf, 1000, "Animation: %d at %lf (dur %lf) S: %f %f %d %d %d %d %d Frame: %lf %lf %lf %lf %f %f %f %d",
 				isSettings_i, time, duration,
